Rename DiagnosticsMonitor to EventPipeSessionManager. Add null checks to EventTracePipeline ctor.
